Fix global visible bookmarks.
The functions of bm.el assume that they are in the same buffer as the bookmark on which they are operating. This means that Anything needs to include the buffer name with the candidates and switch to the buffer containing the selected bookmark (if only temporarily) for bm.el's functions to work correctly.

Signed-off-by: Dan Hackney <dan@haxney.org>